Refactor admin list filters
* Migrate NullableTimestampFilter from the importer
  into the main app
* Add NullableTimestampFilter subclasses for our
  Transcription status fields
