+++ importer/admin.py
@@ -3,6 +3,7 @@ from django.contrib.humanize.templatetags.humanize import naturaltime
 
 from .models import ImportItem, ImportItemAsset, ImportJob
 from .tasks import download_asset_task
+from concordia.admin.filters import NullableTimestampFilter
 
 
 def retry_download_task(modeladmin, request, queryset):
@@ -17,31 +18,6 @@ def retry_download_task(modeladmin, request, queryset):
 
 
 retry_download_task.short_description = "Retry import"
-
-
-class NullableTimestampFilter(admin.SimpleListFilter):
-    """
-    Base class for Admin list filters which define whether a datetime field has
-    a value or is null
-    """
-
-    # Title displayed on the list filter URL
-    title = ""
-    # Model field name:
-    parameter_name = ""
-    # Choices displayed
-    lookup_labels = ("NULL", "NOT NULL")
-
-    def lookups(self, request, model_admin):
-        return zip(("null", "not-null"), self.lookup_labels)
-
-    def queryset(self, request, queryset):
-        kwargs = {"%s__isnull" % self.parameter_name: True}
-        if self.value() == "null":
-            return queryset.filter(**kwargs)
-        elif self.value() == "not-null":
-            return queryset.exclude(**kwargs)
-        return queryset
 
 
 class LastStartedFilter(NullableTimestampFilter):
